twiny_backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
llama-cpp-python
